+++ tests/messaging/test_simple_event_publisher.py
@@ -1,7 +1,6 @@
 from typing import List
 from typing import Union
 
-from dataclasses import dataclass
 from injector import ClassProvider
 from injector import singleton
 
@@ -69,6 +68,12 @@ class EventHandlers:
             self.result['handlers'].append(8)
             self.result['x'] += event.x
 
+    @event_handler
+    def handler9(self, events: list[Event1 | Event3]):
+        for event in events:
+            self.result['handlers'].append(9)
+            self.result['x'] += event.x
+
 
 def test_simple_event_publisher():
     injector = get_injector()
@@ -87,23 +92,23 @@ def test_simple_event_publisher():
     # Assert
     # emit event
     result = event_handlers.result
-    assert result == {'handlers': [1, 2, 7, 8], 'x': 40}
+    assert result == {'handlers': [1, 2, 7, 8, 9], 'x': 50}
 
     # emit same event again
     simple_event_publisher.emit(Event1(20))
-    assert result == {'handlers': [1, 2, 7, 8, 1, 2, 7, 8], 'x': 120}
+    assert result == {'handlers': [1, 2, 7, 8, 9, 1, 2, 7, 8, 9], 'x': 150}
 
     # emit event with no handlers
     simple_event_publisher.emit(Event2(100))
-    assert result == {'handlers': [1, 2, 7, 8, 1, 2, 7, 8], 'x': 120}
+    assert result == {'handlers': [1, 2, 7, 8, 9, 1, 2, 7, 8, 9], 'x': 150}
 
     # emit event with other annotation
     simple_event_publisher.emit(Event3(200))
-    assert result == {'handlers': [1, 2, 7, 8, 1, 2, 7, 8, 3, 7, 8], 'x': 720}
+    assert result == {'handlers': [1, 2, 7, 8, 9, 1, 2, 7, 8, 9, 3, 7, 8, 9], 'x': 950}
 
     # emit event with handler for single event and list of events
     simple_event_publisher.emit(Event4(1000))
-    assert result == {'handlers': [1, 2, 7, 8, 1, 2, 7, 8, 3, 7, 8, 5, 6], 'x': 2720}
+    assert result == {'handlers': [1, 2, 7, 8, 9, 1, 2, 7, 8, 9, 3, 7, 8, 9, 5, 6], 'x': 2950}
 
     # Test emit_many
     event_handlers.result = {'handlers': [], 'x': 0}
@@ -111,13 +116,36 @@ def test_simple_event_publisher():
     # emit event
     simple_event_publisher.emit_many([Event1(10)])
     result = event_handlers.result
-    assert result == {'handlers': [1, 2, 7, 8], 'x': 40}
+    assert result == {'handlers': [1, 2, 7, 8, 9], 'x': 50}
 
     # emit same event twice
     simple_event_publisher.emit_many([Event3(200), Event3(200)])
-    assert result == {'handlers': [1, 2, 7, 8, 3, 3, 7, 7, 8, 8], 'x': 1240}
+    assert result == {'handlers': [1, 2, 7, 8, 9, 3, 3, 7, 7, 8, 8, 9, 9], 'x': 1650}
 
     # emit different events
     simple_event_publisher.emit_many([Event1(10), Event2(100), Event3(200), Event4(1000)])
-    assert result == {'handlers': [1, 2, 7, 8, 3, 3, 7, 7, 8, 8, 1, 2, 7, 7, 8, 8, 3, 5, 6],  'x': 3880}
-
+    assert result == {
+        'handlers': [1, 2,
+                     7,
+                     8,
+                     9,
+                     3,
+                     3,
+                     7,
+                     7,
+                     8,
+                     8,
+                     9,
+                     9,
+                     1,
+                     2,
+                     7,
+                     7,
+                     8,
+                     8,
+                     9,
+                     9,
+                     3,
+                     5,
+                     6], 'x': 4500
+    }